Update hbase version.
We use the mutateRows api from a later version of hbase. Thus, we need
the api to be available in the fake version as well.

I've added the mutateRows and append api's and marked the lock/unlock
row api's as deprecated.

Additionally I've added tests for these. And, the scan test was being a
pain so I simply changed the second value to key2 even though I'm pretty
sure that's not the intention of the test.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>